pyproject.toml
[tool.poetry]
name = "imagehelper"
version = "0.1.0"
description = "Image manipulation"
authors = ["mkshgh"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
Pillow = "^9.0.1"
numpy = "^1.22.3"
python-xmp-toolkit = "^2.0.1"
pyexiv2 = "^2.7.1"
scipy = "^1.8.0"
Wand = "^0.6.7"
matplotlib = "^3.5.1"
selenium = "3.141.0"
PyAutoIt = "^0.6.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
